Update item model - change type double to float on field size

+++ app/models/item.rb
@@ -3,7 +3,7 @@ class Item
   field :name, type: String
   field :description, type: String
   field :size, type: Hash
-  field :price, type: Double
+  field :price, type: Float
   field :item_id, type: Integer
   field :picture, type: String
 end
